# app/src/main/resources/static/swagger-ui/openapi3.yaml
openapi: 3.0.1
info:
  title: GrowIT API Specification
  description: GrowIT description
  version: 0.0.2
servers:
- url: http://growit-alb-alb-549641300.ap-northeast-2.elb.amazonaws.com/
...
paths:
...
  /goals:
    get:
      tags:
      - Goals
      summary: 내 목표 조회
      description: 내 목표 조회
      operationId: get-my-goal
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/auth-signin486549215"
              examples:
                get-my-goal:
                  value: "{\"data\":[]}"
    post:
      tags:
      - Goals
      summary: 목표 생성
      description: 목표 생성
      operationId: create-goal
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/auth-signin486549215"
            examples:
              create-goal:
                value: "{\"name\":\"내 목표는 그로잇 완성\",\"duration\":{\"startDate\":\"\
                  2025-06-23\",\"endDate\":\"2025-07-20\"},\"beforeAfter\":{\"asIs\"\
                  :\"기획 정의\",\"toBe\":\"배포 완료\"},\"plans\":[{\"content\":\"기획 및 설계\
                  \ 회의\"},{\"content\":\"디자인 시안 뽑기\"},{\"content\":\"프론트 개발 및 백 개발\
                  \ 완료\"},{\"content\":\"배포 완료\"}]}"
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/auth-signin486549215"
              examples:
                create-goal:
                  value: "{\"data\":{\"id\":\"2FkZADlb_D6YESeJzr9u3\"}}"
  /goals/{id}:
    delete:
      tags:
      - Goals
      summary: 목표 삭제
      description: 목표 삭제
      operationId: delete-goal
      parameters:
      - name: id
        in: path
        description: ""
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/auth-signin486549215"
              examples:
                delete-goal:
                  value: "{\"data\":\"삭제가 완료 되었습니다.\"}"
...
components:
  schemas:
    auth-signin486549215:
      type: object